fix: Policy violation remediation
The provided Deployment resource violated multiple Kyverno policies. A CUDL patch is applied with the following changes: (1) The AppArmor annotation was removed as it conflicts with seccomp recommendations. (2) The hostPath volume was removed to comply with 'disallow-host-path' and 'restrict-volume-types' policies. (3) The hostPort was removed to comply with the 'disallow-host-ports' policy. (4) The 'privileged' field was set to false and 'allowPrivilegeEscalation' was set to false to comply with 'disallow-privileged-containers' and 'disallow-privilege-escalation' policies. (5) The 'runAsNonRoot' field was added and 'seccompProfile' was set to 'RuntimeDefault' to comply with the 'require-run-as-nonroot' and 'restrict-seccomp-strict' policies. Additionally, the use of dangerous SYS_ADMIN capabilities was eliminated by removing the 'capabilities' section.