fix: Policy violation remediation
The resource definition was non-compliant with multiple Kyverno policies. The AppArmor profile was set to 'runtime/default' and a Seccomp profile annotation was added for POD level policy. HostPath and HostPort usage were removed for compliance with the 'disallow-host-path' and 'disallow-host-ports' policies. The container's privileged access was removed, and 'allowPrivilegeEscalation' was set to false to comply with 'disallow-privileged-containers' and 'disallow-privilege-escalation' policies. The 'SYS_ADMIN' capability was removed in addition to dropping all capabilities initially added, addressing 'disallow-capabilities' policies. Finally, 'runAsNonRoot' was set to true under 'require-run-as-nonroot'. As an improvement suggestion, ensure your container images continue to be updated, locked to a specific version for production environments.